Build: Convert to super build that relies on Lagom shared libraries
Move the CMake rules for the libjs-test262-runner binary into the source
directory so that the top level can be a super build that relies on
ExternalProject_Add for the Lagom dependency. This allows us to build
with a proper find_package(Lagom REQUIRED) build rule, which just feels
warm and fuzzy.

Co-authored-by: sin-ack <sin-ack@users.noreply.github.com>